Moved Asset Registrar trait
