+++ pallets/asset-manager/Cargo.toml
@@ -14,6 +14,7 @@ parallel-primitives = { path = '../../primitives', default-features = false }
 # Substrate
 frame-support 		= { git = 'https://github.com/paritytech/substrate', branch = 'polkadot-v0.9.17', default-features = false }
 frame-system 		= { git = 'https://github.com/paritytech/substrate', branch = 'polkadot-v0.9.17', default-features = false }
+pallet-traits       = { path = '../traits', default-features = false }
 parity-scale-codec 	= { version = '2.0.0', default-features = false, features = [ 'derive' ] }
 scale-info 			= { version = '1.0', default-features = false, features = [ 'derive' ] }
 sp-io 				= { git = 'https://github.com/paritytech/substrate', branch = 'polkadot-v0.9.17', default-features = false }
@@ -36,6 +37,7 @@ std = [
 	'frame-support/std',
 	'frame-system/std',
 	'parity-scale-codec/std',
+	'pallet-traits/std',
 	'scale-info/std',
 	'serde',
 	'sp-io/std',
