--- pallets/traits/src/emergency_filter.rs
@@ -1,3 +1,0 @@
-pub trait EmergencyCallFilter<Call> {
-    fn contains(call: &Call) -> bool;
-}
