+++ runtime/kerria/src/lib.rs
@@ -112,7 +112,7 @@ pub use frame_support::{
     },
     StorageValue,
 };
-use pallet_traits::emergency_filter::EmergencyCallFilter;
+use pallet_traits::EmergencyCallFilter;
 use pallet_xcm::XcmPassthrough;
 #[cfg(any(feature = "std", test))]
 pub use sp_runtime::BuildStorage;
@@ -1242,7 +1242,7 @@ impl Config for XcmConfig {
 // In this case, we tell it to Create an Asset in pallet-assets
 pub struct AssetRegistrar;
 
-impl pallet_asset_manager::AssetRegistrar<Runtime> for AssetRegistrar {
+impl pallet_traits::AssetRegistrar<CurrencyId, Balance, AssetRegistrarMetadata> for AssetRegistrar {
     #[transactional]
     fn create_asset(
         asset: CurrencyId,
